motor and logic class added
